backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic>=2.6.0
llama-index>=0.10.0
langgraph>=0.0.30
neo4j>=5.18.0
qdrant-client>=1.8.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==3.2.2
sqlalchemy>=2.0.29
langchain-groq>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.30
python-dotenv>=1.0.1
networkx>=3.0
playwright>=1.40.0
firecrawl-py>=0.0.20
